Override PYTHONHOME and add "python" command.
If the user already has `PYTHONHOME` set, the binary fails since it cannot
use the python libraries we ship. To overcome this problem, as seen in
ilastik/ilastik#1067, we override the `PYTHONHOME` environment variable.

This patch also adds a "python" argument to `ilastik.bat` that allows
running Python scripts within the same environment ilastik is run.